# feat: Improve performance preventing deepcopy