@hille721 2023-05-18
Hi @andreatomassetti,

good catch! Already thought during implementation that this could scale badly. It is not directly related to the size of the inventory but more to the number of files where a variable of the server could be in (e.g. having a lot of nested groups) and the file size in the end. 

I already tested with a really large inventory from my company, from ~10 seconds to ~5 seconds. Not that significant as for your example, but still an awesome improvement. Thanks!