@andreatomassetti 2023-05-18
> I already tested with a really large inventory from my company, from ~10 seconds to ~5 seconds. Not that significant as for your example, but still an awesome improvement. Thanks!

Yeah we have a really nested inventory, maybe is that more than how big it is. Or maybe it is just my computer that it's slow :joy: